+++ src/interfaces/IConfigCreator.h
@@ -4,8 +4,7 @@
  * Copyright 2014      Lucas Jones <https://github.com/lucasjones>
  * Copyright 2014-2016 Wolf9466    <https://github.com/OhGodAPet>
  * Copyright 2016      Jay D Dee   <jayddee246@gmail.com>
- * Copyright 2017-2018 XMR-Stak    <https://github.com/fireice-uk>, <https://github.com/psychocrypt>
- * Copyright 2016-2018 XMRig       <https://github.com/xmrig>, <support@xmrig.com>
+ * Copyright 2016-2018 XMRig       <support@xmrig.com>
  *
  *   This program is free software: you can redistribute it and/or modify
  *   it under the terms of the GNU General Public License as published by
@@ -21,28 +20,26 @@
  *   along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
 
-
-#include <winsock2.h>
-#include <windows.h>
-
-
-#include "App.h"
-#include "core/Config.h"
-#include "core/Controller.h"
+#ifndef __ICONFIGCREATOR_H__
+#define __ICONFIGCREATOR_H__
 
 
-void App::background()
+namespace xmrig {
+
+
+class IConfig;
+
+
+class IConfigCreator
 {
-    if (!m_controller->config()->isBackground()) {
-        return;
-    }
+public:
+    virtual ~IConfigCreator() {}
 
-    HWND hcon = GetConsoleWindow();
-    if (hcon) {
-        ShowWindow(hcon, SW_HIDE);
-    } else {
-        HANDLE h = GetStdHandle(STD_OUTPUT_HANDLE);
-        CloseHandle(h);
-        FreeConsole();
-    }
-}
+    virtual IConfig *create() const = 0;
+};
+
+
+} /* namespace xmrig */
+
+
+#endif // __ICONFIGCREATOR_H__