+++ src/net/Client.h
@@ -32,8 +32,9 @@
 
 #include "net/Id.h"
 #include "net/Job.h"
+#include "net/Storage.h"
 #include "net/SubmitResult.h"
-#include "net/Url.h"
+#include "net/Pool.h"
 #include "rapidjson/fwd.h"
 
 
@@ -53,31 +54,29 @@ public:
     };
 
     constexpr static int kResponseTimeout  = 20 * 1000;
-    constexpr static int kKeepAliveTimeout = 60 * 1000;
 
     Client(int id, const char *agent, IClientListener *listener);
+    ~Client();
 
     bool disconnect();
     int64_t submit(const JobResult &result);
     void connect();
-    void connect(const Url *url);
+    void connect(const Pool &pool);
     void deleteLater();
-    void setUrl(const Url *url);
+    void setPool(const Pool &pool);
     void tick(uint64_t now);
 
     inline bool isReady() const              { return m_state == ConnectedState && m_failures == 0; }
-    inline const char *host() const          { return m_url.host(); }
+    inline const char *host() const          { return m_pool.host(); }
     inline const char *ip() const            { return m_ip; }
     inline const Job &job() const            { return m_job; }
     inline int id() const                    { return m_id; }
     inline SocketState state() const         { return m_state; }
-    inline uint16_t port() const             { return m_url.port(); }
+    inline uint16_t port() const             { return m_pool.port(); }
     inline void setQuiet(bool quiet)         { m_quiet = quiet; }
     inline void setRetryPause(int ms)        { m_retryPause = ms; }
 
 private:
-    ~Client();
-
     bool close();
     bool isCriticalError(const char *message);
     bool parseJob(const rapidjson::Value &params, int *code);
@@ -103,7 +102,7 @@ private:
     static void onRead(uv_stream_t *stream, ssize_t nread, const uv_buf_t *buf);
     static void onResolved(uv_getaddrinfo_t *req, int status, struct addrinfo *res);
 
-    static inline Client *getClient(void *data) { return static_cast<Client*>(data); }
+    static inline Client *getClient(void *data) { return m_storage.get(data); }
 
     addrinfo m_hints;
     bool m_ipv6;
@@ -118,22 +117,22 @@ private:
     int m_retryPause;
     int64_t m_failures;
     Job m_job;
+    Pool m_pool;
     size_t m_recvBufPos;
     SocketState m_state;
-    static int64_t m_sequence;
     std::map<int64_t, SubmitResult> m_results;
     uint64_t m_expire;
     uint64_t m_jobs;
-    Url m_url;
+    uint64_t m_keepAlive;
+    uintptr_t m_key;
     uv_buf_t m_recvBuf;
     uv_getaddrinfo_t m_resolver;
     uv_stream_t *m_stream;
     uv_tcp_t *m_socket;
     xmrig::Id m_rpcId;
 
-#   ifndef XMRIG_PROXY_PROJECT
-    uv_timer_t m_keepAliveTimer;
-#   endif
+    static int64_t m_sequence;
+    static xmrig::Storage<Client> m_storage;
 };