+++ src/proxy/splitters/nicehash/NonceMapper.cpp
@@ -34,7 +34,6 @@
 #include "net/strategies/DonateStrategy.h"
 #include "net/strategies/FailoverStrategy.h"
 #include "net/strategies/SinglePoolStrategy.h"
-#include "net/Url.h"
 #include "proxy/Counters.h"
 #include "proxy/Error.h"
 #include "proxy/events/AcceptEvent.h"
@@ -106,7 +105,7 @@ void NonceMapper::gc()
 }
 
 
-void NonceMapper::reload(const std::vector<Url*> &pools)
+void NonceMapper::reload(const std::vector<Pool> &pools)
 {
     delete m_pending;
 
@@ -183,7 +182,7 @@ void NonceMapper::onActive(IStrategy *strategy, Client *client)
         m_pending  = nullptr;
     }
 
-    if (m_controller->config()->verbose()) {
+    if (m_controller->config()->isVerbose()) {
         LOG_INFO(isColors() ? "#%03u \x1B[01;37muse pool \x1B[01;36m%s:%d \x1B[01;30m%s" : "#%03u use pool %s:%d %s",
                  m_id, client->host(), client->port(), client->ip());
     }
@@ -192,7 +191,7 @@ void NonceMapper::onActive(IStrategy *strategy, Client *client)
 
 void NonceMapper::onJob(IStrategy *strategy, Client *client, const Job &job)
 {
-    if (m_controller->config()->verbose()) {
+    if (m_controller->config()->isVerbose()) {
         LOG_INFO(isColors() ? "#%03u \x1B[01;35mnew job\x1B[0m from \x1B[01;37m%s:%d\x1B[0m diff \x1B[01;37m%d" : "#%03u new job from %s:%d diff %d",
                  m_id, client->host(), client->port(), job.diff());
     }
@@ -236,11 +235,11 @@ void NonceMapper::onResultAccepted(IStrategy *strategy, Client *client, const Su
 
 bool NonceMapper::isColors() const
 {
-    return m_controller->config()->colors();
+    return m_controller->config()->isColors();
 }
 
 
-IStrategy *NonceMapper::createStrategy(const std::vector<Url*> &pools)
+IStrategy *NonceMapper::createStrategy(const std::vector<Pool> &pools)
 {
     if (pools.size() > 1) {
         return new FailoverStrategy(pools, m_controller->config()->retryPause(), m_controller->config()->retries(), this);