+++ src/proxy/splitters/simple/SimpleMapper.h
@@ -32,6 +32,7 @@
 
 #include "interfaces/IStrategyListener.h"
 #include "net/Job.h"
+#include "net/Pool.h"
 
 
 class DonateStrategy;
@@ -57,9 +58,10 @@ public:
     ~SimpleMapper();
 
     void add(Miner *miner, const LoginRequest &request);
-    void reload(const std::vector<Url*> &pools);
+    void reload(const std::vector<Pool> &pools);
     void remove(const Miner *miner);
     void reuse(Miner *miner, const LoginRequest &request);
+    void stop();
     void submit(SubmitEvent *event);
     void tick(uint64_t ticks, uint64_t now);
 
@@ -77,7 +79,7 @@ protected:
 private:
     bool isColors() const;
     bool isValidJobId(const xmrig::Id &id) const;
-    IStrategy *createStrategy(const std::vector<Url*> &pools);
+    IStrategy *createStrategy(const std::vector<Pool> &pools);
     void connect();
     void setJob(const Job &job);